Update README.md
This repository is for educational purposes only. We do not condone or encourage any illegal activity. Unauthorized use of ransomware or any form of hacking is illegal and punishable by law. Always practice ethical hacking and use your knowledge responsibly. The contributors of this repository are not liable for any misuse or damages resulting from the code provided.